catalog-implementation
Ignore-this: c5358db293c7a859f4f8755c220c7ab9

darcs-hash:20090603172135-68901-57e07200caa58df9d09b7351d2ded36fa4bd64bb.gz

+++ kernel/tr/pstr/pstr_long.h
@@ -202,7 +202,7 @@ public:
 		//no error checks here, TODO - add some assertions
 		//we don't care if we go before the start of string too far
 		--m_pos;
-		if (--ofs < PSTR_LONG_BLK_HDR_SIZE)
+		if (((int)(--ofs)) < PSTR_LONG_BLK_HDR_SIZE)
 		{
 			CHECKP(blk);
 			if (PSTR_LONG_BLK_HDR(blk)->prev_blk != XNULL)
